requirements.txt
webdriver_manager
selenium==4.6.0
pygame
pyyaml